Recover from structurally invalid snapshots
