+++ CMakeLists.txt
@@ -1126,12 +1126,6 @@ if(BUILD_TESTS)
   )
 
   add_e2e_test(
-    NAME recovery_corrupt_snapshot_test
-    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/recovery_corrupt_snapshot.py
-    BUCKET bucket_b
-  )
-
-  add_e2e_test(
     NAME recovery_test_suite
     PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/e2e_suite.py
     LABEL suite
