+++ tests/e2e_operations.py
@@ -817,36 +817,48 @@ def test_nulled_snapshot(network, args):
     with tempfile.TemporaryDirectory() as snapshots_dir:
         LOG.debug(f"Using {snapshots_dir} as snapshots directory")
 
-        snapshot_name = find_latest_committed_snapshot_name(network)
-
-        with open(
-            os.path.join(snapshots_dir, snapshot_name), "wb+"
-        ) as temp_empty_snapshot:
-
-            LOG.debug(f"Created empty snapshot {temp_empty_snapshot.name}")
-            temp_empty_snapshot.write(b"\x00" * 64)
-
-        LOG.info(
-            "Attempt to join a node using the corrupted snapshot copy (should fail)"
+        valid_snapshot_name, corrupt_snapshot_name = (
+            find_latest_committed_snapshot_name(network, count=2)
         )
-        new_node = network.create_node()
-        failed = False
-        try:
-            network.join_node(
-                new_node,
-                args.package,
-                args,
-                snapshots_dir=snapshots_dir,
-                # Don't try to fetch a snapshot, look at the local files
-                fetch_recent_snapshot=False,
-                from_snapshot=True,
-            )
-        except Exception as e:
-            failed = True
-            LOG.info(f"Node failed to join as expected: {e}")
+        primary, _ = network.find_primary()
+        committed_snapshots_dir = network.get_committed_snapshots(primary)
+        shutil.copy(
+            os.path.join(committed_snapshots_dir, valid_snapshot_name), snapshots_dir
+        )
 
-        # (Existing assertion logic retained)
-        assert failed, "Node should not have joined successfully"
+        corrupt_snapshot_path = os.path.join(snapshots_dir, corrupt_snapshot_name)
+        with open(corrupt_snapshot_path, "wb+") as corrupt_snapshot:
+            LOG.debug(f"Created corrupt snapshot {corrupt_snapshot.name}")
+            corrupt_snapshot.write(b"\x00" * 64)
+
+        LOG.info("Join from the older valid snapshot")
+        new_node = network.create_node()
+        network.join_node(
+            new_node,
+            args.package,
+            args,
+            snapshots_dir=snapshots_dir,
+            # Don't try to fetch a snapshot, look at the local files
+            fetch_recent_snapshot=False,
+            from_snapshot=True,
+        )
+        new_node.stop()
+
+        assert new_node.check_log_for_error_message(
+            f"{corrupt_snapshot_name} cannot be parsed"
+        )
+        valid_snapshot_seqno = ccf.ledger.snapshot_index_from_filename(
+            valid_snapshot_name
+        )
+        assert new_node.check_log_for_error_message(
+            f"Setting startup snapshot seqno to {valid_snapshot_seqno}"
+        )
+
+        node_snapshots_dir = os.path.join(
+            new_node.remote.remote.root,
+            new_node.remote.snapshots_dir_name,
+        )
+        os.remove(os.path.join(node_snapshots_dir, corrupt_snapshot_name))
 
 
 def test_corrupt_snapshot_handling(network, args):
